requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0
